+++ backend/src/main/java/org/cryptomator/hub/api/GroupDto.java
@@ -5,12 +5,20 @@ import org.cryptomator.hub.entities.Group;
 
 public final class GroupDto extends AuthorityDto {
 
-	GroupDto(@JsonProperty("id") String id, @JsonProperty("name") String name) {
+	@JsonProperty("memberSize")
+	public final Integer memberSize;
+
+	GroupDto(@JsonProperty("id") String id, @JsonProperty("name") String name, @JsonProperty("memberSize") Integer memberSize) {
 		super(id, Type.GROUP, name, null);
+		this.memberSize = memberSize;
 	}
 
 	public static GroupDto fromEntity(Group group) {
-		return new GroupDto(group.getId(), group.getName());
+		return fromEntity(group, false);
 	}
 
+	public static GroupDto fromEntity(Group group, boolean withMemberSize) {
+		Integer memberSize = withMemberSize ? group.getMemberSize() : null;
+		return new GroupDto(group.getId(), group.getName(), memberSize);
+	}
 }